+++ src/components/learnDetail/memo/MemoDotButton.tsx
@@ -1,8 +1,7 @@
 import { ImageDiv } from '@src/components/common';
-import { MemoConfirmModalKey } from '@src/components/learnDetail/modal/ConfirmModal';
 import { useClickOutside } from '@src/hooks/common';
-import { MemoState } from '@src/pages/learn/[id]';
-import { MemoData } from '@src/types/learn-detail';
+import { MemoConfirmModalKey, MemoState } from '@src/types/learnDetail';
+import { MemoData } from '@src/types/learnDetail/remote';
 import dynamic from 'next/dynamic';
 import { icDotDefault, icDotHover } from 'public/assets/icons';
 import { Dispatch, SetStateAction, useRef, useState } from 'react';
