+++ src/components/learnDetail/memo/MemoDropdown.tsx
@@ -1,9 +1,8 @@
-import { MemoConfirmModalKey } from '@src/components/learnDetail/modal/ConfirmModal';
-import { MemoState } from '@src/pages/learn/[id]';
-import { MemoData } from '@src/types/learn-detail';
+import { MemoData } from '@src/types/learnDetail/remote';
 import { COLOR, FONT_STYLES } from '@src/styles';
 import { Dispatch, SetStateAction } from 'react';
 import styled from 'styled-components';
+import { MemoState, MemoConfirmModalKey } from '@src/types/learnDetail';
 
 interface MemoDropdownProps {
   memoData: MemoData;
