+++ src/components/learnDetail/memo/MemoItem.tsx
@@ -1,9 +1,8 @@
 import { ImageDiv } from '@src/components/common';
 import { MemoDotButton, MemoForm } from '@src/components/learnDetail/memo';
-import { MemoConfirmModalKey } from '@src/components/learnDetail/modal/ConfirmModal';
-import { MemoState } from '@src/pages/learn/[id]';
-import { MemoData } from '@src/types/learn-detail';
 import { COLOR, FONT_STYLES } from '@src/styles';
+import { MemoConfirmModalKey, MemoState } from '@src/types/learnDetail';
+import { MemoData } from '@src/types/learnDetail/remote';
 import { MEMO_CONTENT_MAX } from '@src/utils/constant';
 import { useRouter } from 'next/router';
 import { icArrowUp } from 'public/assets/icons';
