+++ src/components/learnDetail/memo/MemoLog.tsx
@@ -1,8 +1,7 @@
 import { EmptyMemo, MemoItem } from '@src/components/learnDetail/memo';
-import { MemoConfirmModalKey } from '@src/components/learnDetail/modal/ConfirmModal';
-import { MemoState } from '@src/pages/learn/[id]';
-import { MemoData } from '@src/types/learn-detail';
 import { COLOR } from '@src/styles';
+import { MemoState, MemoConfirmModalKey } from '@src/types/learnDetail';
+import { MemoData } from '@src/types/learnDetail/remote';
 import { Dispatch, SetStateAction } from 'react';
 import styled from 'styled-components';
