+++ src/hooks/learnDetail/useUpdateMemoList.ts
@@ -1,8 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { MemoConfirmModalKey } from '@src/components/learnDetail/modal/ConfirmModal';
-import { MemoInfo, MemoState } from '@src/pages/learn/[id]';
 import { api } from '@src/services/api';
-import { MemoData } from '@src/types/learn-detail';
+import { MemoState, MemoInfo, MemoConfirmModalKey } from '@src/types/learnDetail';
+import { MemoData } from '@src/types/learnDetail/remote';
 import { INITIAL_MEMO_STATE } from '@src/utils/constant';
 
 interface useUpdateMemoListProps {
