+++ src/pages/learn/index.tsx
@@ -1,7 +1,7 @@
 import { Footer, ImageDiv, NavigationBar, NewsList, Pagination, SEO, VideoListSkeleton } from '@src/components/common';
 import { SelectBox } from '@src/components/learn';
 import { api } from '@src/services/api';
-import { PostSearchConditionRequestBody, VideoData } from '@src/types/learn';
+import { PostSearchConditionRequestBody, VideoData } from '@src/types/learn/remote';
 import { loginState } from '@src/stores/loginState';
 import { COLOR, FONT_STYLES } from '@src/styles';
 import { BLOCK_SIZE, categoryList, channelList, LIST_SIZE, speakerList } from '@src/utils/constant';
