+++ src/services/api/home.ts
@@ -1,4 +1,4 @@
-import { VideoListData } from '@src/types/home';
+import { VideoListData } from '@src/types/home/remote';
 
 export interface HomeService {
   getVideoData(): Promise<VideoListData>;
