+++ src/services/api/learn.ts
@@ -1,4 +1,4 @@
-import { PostSearchConditionRequestBody, PostSearchConditionResponse } from '@src/types/learn';
+import { PostSearchConditionRequestBody, PostSearchConditionResponse } from '@src/types/learn/remote';
 
 export interface LearnService {
   postSearchCondition(body: PostSearchConditionRequestBody): Promise<PostSearchConditionResponse>;
