+++ src/services/api/review.ts
@@ -1,4 +1,4 @@
-import { PostReviewRequestBody, PostReviewResponse, ReviewTab } from '@src/types/review';
+import { PostReviewRequestBody, PostReviewResponse, ReviewTab } from '@src/types/review/remote';
 
 export interface ReviewService {
   postReviewVideoList(body: PostReviewRequestBody, reviewTab: ReviewTab): Promise<PostReviewResponse>;
