+++ src/types/learnDetail/remote.ts
